feat: pre-generate RFP summaries via backfill script
